+++ MonoTanksBotLogic/Networking/Packets/gameState/Tiles/OwnTurret.cs
@@ -8,4 +8,4 @@ namespace MonoTanksClientLogic.Models;
 /// <param name="Direction">Represents turret direction.</param>
 /// <param name="BulletCount">Represents number of available bullets.</param>
 /// /// <param name="TicksToRegenBullet">Represents time in ticks to regenerate bullet.</param>
-public record class OwnTurret(Direction Direction, long BulletCount, double? TicksToRegenBullet);
+public record class OwnTurret(Direction Direction, int BulletCount, double? TicksToRegenBullet);
